Fixed some logic issues in the PackageSelection() and FindChocoPackage() functions.
Added launch.json configuration for running script with -e option.
When executed from install.bat, the installer.ps1 script utilizes a -e option
to execute from the temp directory. Otherwise the script will execute from whatever local directory it is ran out of.
This is so the script can be run independently of the install.bat script.

PackageReference is updated with correlating packages for quicker lookup times.